\name{detergent}
\docType{data}
\alias{detergent}
\title{Detergent Brand Choice}
\description{
  This data set gives the detergent brand choice by households and the
  price of each brand. 
}
\usage{data(detergent)}
\format{A matrix containing 7 variables and 2657 observations}
\references{Chintagunta, P. K. and Prasad, A. R. (1998) \dQuote{An Empirical
  Investigation of the ``Dynamic McFadden'' Model of Purchase Timing and
  Brand Choice: Implications for Market Structure}. \emph{Journal of Business and
  Economic Statistics} vol. 16 no. 1 pp.2-12.}
\keyword{datasets}
