\name{japan}
\docType{data}
\alias{japan}
\title{Voters' Preferences of Political Parties in Japan (1995)}
\description{
  This dataset gives voters' preferences of political parties (Liberal
  Democratic Party (LDP), New Frontier Party (NFP), Sakigake (SKG), and
  Japanese Communist Party (JCP)) in Japan on the 0 (least preferred) - 100
  (most preferred) scale. It is based on the 1995 survey data of 418
  individual voters. The data also include the sex, education and age of
  those voters. 
}

\usage{data(japan)}

\format{A matrix containing the following 7 variables for 418 observations.
  
  \tabular{lll}{
    LDP \tab preference for Liberal Democratic Party \tab 0 - 100 \cr
    NFP \tab preference for New Frontier Party \tab 0 - 100 \cr
    SKG \tab preference for Sakigake \tab 0 - 100 \cr
    JCP \tab preference for Japanse Communist Party \tab 0 - 100 \cr
    sex \tab sex of each voter \tab 1 = male, 2 = female \cr
    education \tab levels of education for each voter \tab \cr
    age \tab age of each voter \tab
  }
}
\keyword{datasets}
