% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mnp.R
\name{print.summary.mnp}
\alias{print.summary.mnp}
\title{Print the summary of the results for the Multinomial Probit Models}
\usage{
\method{print}{summary.mnp}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class \code{summary.mnp}.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{summary} print method for class \code{mnp}.
}
\seealso{
\code{mnp}; MNP home page at
\url{http://imai.princeton.edu/research/MNP.html}
}
\author{
Kosuke Imai, Department of Politics, Princeton University
\email{kimai@Princeton.Edu}
}
\keyword{methods}
