\name{detergent}
\docType{data}
\alias{detergent}
\title{Detergent Brand Choice}
\description{
  This data set gives the detergent brand choice by households and the
  price of each brand. 
}
\usage{data(detergent)}

\format{A matrix containing the following 7 variables and 2657 observations.

  \tabular{lll}{
    choice \tab factor \tab a brand chosen by each household\cr
    Tide \tab numeric \tab log price of Tide\cr
    Wisk \tab numeric \tab log price of Wisk\cr
    EraPlus \tab numeric \tab log price of EraPlus\cr
    Surf \tab numeric \tab log price of Surf\cr
    Solo \tab numeric \tab log price of Solo\cr
    All \tab numeric \tab log price of All\cr
  }

}



\references{Chintagunta, P. K. and Prasad, A. R. (1998) \dQuote{An Empirical
  Investigation of the ``Dynamic McFadden'' Model of Purchase Timing and
  Brand Choice: Implications for Market Structure}. \emph{Journal of Business and
  Economic Statistics} vol. 16 no. 1 pp.2-12.}
\keyword{datasets}
