\name{japan}
\docType{data}
\alias{japan}
\title{Voters' Preferences of Political Parties in Japan (1995)}
\description{
  This data set gives voters' preferences of political parties (Liberal
  Democratic Party (LDP), New Frontier Party (NFP), Sakigake (SKG), and
  Japanese Communist Party (JCP)) in Japan. It is based on the 1995
  survey data. The data also include the sex, education and age of
  respondents. 
}
\usage{data(japan)}
\format{A matrix containing 7 variables and 418 observations}
\keyword{datasets}
