% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japan.R
\docType{data}
\name{japan}
\alias{japan}
\title{Voters' Preferences of Political Parties in Japan (1995)}
\format{
A data frame containing the following 7 variables for 418
observations.

\tabular{llll}{ LDP \tab numeric \tab preference for Liberal Democratic
Party \tab 0 - 100 \cr NFP \tab numeric \tab preference for New Frontier
Party \tab 0 - 100 \cr SKG \tab numeric \tab preference for Sakigake \tab 0
- 100 \cr JCP \tab numeric \tab preference for Japanese Communist Party \tab
0 - 100 \cr gender \tab factor \tab gender of each voter \tab \code{male} or
\code{female} \cr education \tab numeric \tab levels of education for each
voter \tab \cr age \tab numeric \tab age of each voter \tab }
}
\description{
This dataset gives voters' preferences of political parties in Japan on the
0 (least preferred) - 100 (most preferred) scale.  It is based on the 1995
survey data of 418 individual voters.  The data also include the sex,
education level, and age of the voters. The survey allowed voters to choose
among four parties: Liberal Democratic Party (LDP), New Frontier Party
(NFP), Sakigake (SKG), and Japanese Communist Party (JCP).
}
\keyword{datasets}
